[SYNC][V1.5][FAIL] optimized_octree.hpp: attempt to regroup allocations
This version attempt to replace the "new struct node" by a vector of "struct node"
with indexes instead of pointer. The result is underwhelming and suffer a huge slow
as soon as we hit 800k items. A dynamic_bitset have been created to locate the
existing nodes, but since we don"t destroy nodes, it has been replaced by an index
in the implementation