[SYNC][V1][FIX] optimized_octree.hpp: clean, optimized and fix
Everything have be re-thinked in an optimized way and some fixes have been made.
The tests now pass in the case of N == 1 and N == 2 (don't know why for the others).
The function under FIXME is the only thing left for v1 and should be done asap,
since it will bring way better performance (search takes 1/3s, way too long).